swc_ecma_regexp::ast

Enum Term

pub enum Term {
    BoundaryAssertion(Box<BoundaryAssertion>),
    LookAroundAssertion(Box<LookAroundAssertion>),
    Quantifier(Box<Quantifier>),
    Character(Box<Character>),
    Dot(Dot),
    CharacterClassEscape(Box<CharacterClassEscape>),
    UnicodePropertyEscape(Box<UnicodePropertyEscape>),
    CharacterClass(Box<CharacterClass>),
    CapturingGroup(Box<CapturingGroup>),
    IgnoreGroup(Box<IgnoreGroup>),
    IndexedReference(Box<IndexedReference>),
    NamedReference(Box<NamedReference>),
}

Single unit of Alternative, containing various kinds.

Variants

BoundaryAssertion(Box<BoundaryAssertion>)

LookAroundAssertion(Box<LookAroundAssertion>)

Quantifier(Box<Quantifier>)

Character(Box<Character>)

Dot(Dot)

CharacterClassEscape(Box<CharacterClassEscape>)

UnicodePropertyEscape(Box<UnicodePropertyEscape>)

CharacterClass(Box<CharacterClass>)

CapturingGroup(Box<CapturingGroup>)

IgnoreGroup(Box<IgnoreGroup>)

IndexedReference(Box<IndexedReference>)

NamedReference(Box<NamedReference>)

Implementations

impl Term

pub const fn is_boundary_assertion(&self) -> bool

Returns true if self is of variant BoundaryAssertion.

pub fn as_boundary_assertion(&self) -> Option<&Box<BoundaryAssertion>>

Returns Some if self is a reference of variant BoundaryAssertion, and None otherwise.

pub fn as_mut_boundary_assertion( &mut self, ) -> Option<&mut Box<BoundaryAssertion>>

Returns Some if self is a mutable reference of variant BoundaryAssertion, and None otherwise.

pub fn expect_boundary_assertion(self) -> Box<BoundaryAssertion>
where Term: Debug,

Unwraps the value, yielding the content of BoundaryAssertion.

Panics

Panics if the value is not BoundaryAssertion. In debug builds the panic message includes the content of self.

pub fn boundary_assertion(self) -> Option<Box<BoundaryAssertion>>

Returns Some if self is of variant BoundaryAssertion, and None otherwise.

pub const fn is_look_around_assertion(&self) -> bool

Returns true if self is of variant LookAroundAssertion.

pub fn as_look_around_assertion(&self) -> Option<&Box<LookAroundAssertion>>

Returns Some if self is a reference of variant LookAroundAssertion, and None otherwise.

pub fn as_mut_look_around_assertion( &mut self, ) -> Option<&mut Box<LookAroundAssertion>>

Returns Some if self is a mutable reference of variant LookAroundAssertion, and None otherwise.

pub fn expect_look_around_assertion(self) -> Box<LookAroundAssertion>
where Term: Debug,

Unwraps the value, yielding the content of LookAroundAssertion.

Panics

Panics if the value is not LookAroundAssertion. In debug builds the panic message includes the content of self.

pub fn look_around_assertion(self) -> Option<Box<LookAroundAssertion>>

Returns Some if self is of variant LookAroundAssertion, and None otherwise.

pub const fn is_quantifier(&self) -> bool

Returns true if self is of variant Quantifier.

pub fn as_quantifier(&self) -> Option<&Box<Quantifier>>

Returns Some if self is a reference of variant Quantifier, and None otherwise.

pub fn as_mut_quantifier(&mut self) -> Option<&mut Box<Quantifier>>

Returns Some if self is a mutable reference of variant Quantifier, and None otherwise.

pub fn expect_quantifier(self) -> Box<Quantifier>
where Term: Debug,

Unwraps the value, yielding the content of Quantifier.

Panics

Panics if the value is not Quantifier. In debug builds the panic message includes the content of self.

pub fn quantifier(self) -> Option<Box<Quantifier>>

Returns Some if self is of variant Quantifier, and None otherwise.

pub const fn is_character(&self) -> bool

Returns true if self is of variant Character.

pub fn as_character(&self) -> Option<&Box<Character>>

Returns Some if self is a reference of variant Character, and None otherwise.

pub fn as_mut_character(&mut self) -> Option<&mut Box<Character>>

Returns Some if self is a mutable reference of variant Character, and None otherwise.

pub fn expect_character(self) -> Box<Character>
where Term: Debug,

Unwraps the value, yielding the content of Character.

Panics

Panics if the value is not Character. In debug builds the panic message includes the content of self.

pub fn character(self) -> Option<Box<Character>>

Returns Some if self is of variant Character, and None otherwise.

pub const fn is_dot(&self) -> bool

Returns true if self is of variant Dot.

pub fn as_dot(&self) -> Option<&Dot>

Returns Some if self is a reference of variant Dot, and None otherwise.

pub fn as_mut_dot(&mut self) -> Option<&mut Dot>

Returns Some if self is a mutable reference of variant Dot, and None otherwise.

pub fn expect_dot(self) -> Dot
where Term: Debug,

Unwraps the value, yielding the content of Dot.

Panics

Panics if the value is not Dot. In debug builds the panic message includes the content of self.

pub fn dot(self) -> Option<Dot>

Returns Some if self is of variant Dot, and None otherwise.

pub const fn is_character_class_escape(&self) -> bool

Returns true if self is of variant CharacterClassEscape.

pub fn as_character_class_escape(&self) -> Option<&Box<CharacterClassEscape>>

Returns Some if self is a reference of variant CharacterClassEscape, and None otherwise.

pub fn as_mut_character_class_escape( &mut self, ) -> Option<&mut Box<CharacterClassEscape>>

Returns Some if self is a mutable reference of variant CharacterClassEscape, and None otherwise.

pub fn expect_character_class_escape(self) -> Box<CharacterClassEscape>
where Term: Debug,

Unwraps the value, yielding the content of CharacterClassEscape.

Panics

Panics if the value is not CharacterClassEscape. In debug builds the panic message includes the content of self.

pub fn character_class_escape(self) -> Option<Box<CharacterClassEscape>>

Returns Some if self is of variant CharacterClassEscape, and None otherwise.

pub const fn is_unicode_property_escape(&self) -> bool

Returns true if self is of variant UnicodePropertyEscape.

pub fn as_unicode_property_escape(&self) -> Option<&Box<UnicodePropertyEscape>>

Returns Some if self is a reference of variant UnicodePropertyEscape, and None otherwise.

pub fn as_mut_unicode_property_escape( &mut self, ) -> Option<&mut Box<UnicodePropertyEscape>>

Returns Some if self is a mutable reference of variant UnicodePropertyEscape, and None otherwise.

pub fn expect_unicode_property_escape(self) -> Box<UnicodePropertyEscape>
where Term: Debug,

Unwraps the value, yielding the content of UnicodePropertyEscape.

Panics

Panics if the value is not UnicodePropertyEscape. In debug builds the panic message includes the content of self.

pub fn unicode_property_escape(self) -> Option<Box<UnicodePropertyEscape>>

Returns Some if self is of variant UnicodePropertyEscape, and None otherwise.

pub const fn is_character_class(&self) -> bool

Returns true if self is of variant CharacterClass.

pub fn as_character_class(&self) -> Option<&Box<CharacterClass>>

Returns Some if self is a reference of variant CharacterClass, and None otherwise.

pub fn as_mut_character_class(&mut self) -> Option<&mut Box<CharacterClass>>

Returns Some if self is a mutable reference of variant CharacterClass, and None otherwise.

pub fn expect_character_class(self) -> Box<CharacterClass>
where Term: Debug,

Unwraps the value, yielding the content of CharacterClass.

Panics

Panics if the value is not CharacterClass. In debug builds the panic message includes the content of self.

pub fn character_class(self) -> Option<Box<CharacterClass>>

Returns Some if self is of variant CharacterClass, and None otherwise.

pub const fn is_capturing_group(&self) -> bool

Returns true if self is of variant CapturingGroup.

pub fn as_capturing_group(&self) -> Option<&Box<CapturingGroup>>

Returns Some if self is a reference of variant CapturingGroup, and None otherwise.

pub fn as_mut_capturing_group(&mut self) -> Option<&mut Box<CapturingGroup>>

Returns Some if self is a mutable reference of variant CapturingGroup, and None otherwise.

pub fn expect_capturing_group(self) -> Box<CapturingGroup>
where Term: Debug,

Unwraps the value, yielding the content of CapturingGroup.

Panics

Panics if the value is not CapturingGroup. In debug builds the panic message includes the content of self.

pub fn capturing_group(self) -> Option<Box<CapturingGroup>>

Returns Some if self is of variant CapturingGroup, and None otherwise.

pub const fn is_ignore_group(&self) -> bool

Returns true if self is of variant IgnoreGroup.

pub fn as_ignore_group(&self) -> Option<&Box<IgnoreGroup>>

Returns Some if self is a reference of variant IgnoreGroup, and None otherwise.

pub fn as_mut_ignore_group(&mut self) -> Option<&mut Box<IgnoreGroup>>

Returns Some if self is a mutable reference of variant IgnoreGroup, and None otherwise.

pub fn expect_ignore_group(self) -> Box<IgnoreGroup>
where Term: Debug,

Unwraps the value, yielding the content of IgnoreGroup.

Panics

Panics if the value is not IgnoreGroup. In debug builds the panic message includes the content of self.

pub fn ignore_group(self) -> Option<Box<IgnoreGroup>>

Returns Some if self is of variant IgnoreGroup, and None otherwise.

pub const fn is_indexed_reference(&self) -> bool

Returns true if self is of variant IndexedReference.

pub fn as_indexed_reference(&self) -> Option<&Box<IndexedReference>>

Returns Some if self is a reference of variant IndexedReference, and None otherwise.

pub fn as_mut_indexed_reference(&mut self) -> Option<&mut Box<IndexedReference>>

Returns Some if self is a mutable reference of variant IndexedReference, and None otherwise.

pub fn expect_indexed_reference(self) -> Box<IndexedReference>
where Term: Debug,

Unwraps the value, yielding the content of IndexedReference.

Panics

Panics if the value is not IndexedReference. In debug builds the panic message includes the content of self.

pub fn indexed_reference(self) -> Option<Box<IndexedReference>>

Returns Some if self is of variant IndexedReference, and None otherwise.

pub const fn is_named_reference(&self) -> bool

Returns true if self is of variant NamedReference.

pub fn as_named_reference(&self) -> Option<&Box<NamedReference>>

Returns Some if self is a reference of variant NamedReference, and None otherwise.

pub fn as_mut_named_reference(&mut self) -> Option<&mut Box<NamedReference>>

Returns Some if self is a mutable reference of variant NamedReference, and None otherwise.

pub fn expect_named_reference(self) -> Box<NamedReference>
where Term: Debug,

Unwraps the value, yielding the content of NamedReference.

Panics

Panics if the value is not NamedReference. In debug builds the panic message includes the content of self.

pub fn named_reference(self) -> Option<Box<NamedReference>>

Returns Some if self is of variant NamedReference, and None otherwise.